+++ src/test/java/com/configcat/AutoPollingPolicyTest.java
@@ -1,5 +1,8 @@
 package com.configcat;
 
+import com.configcat.cache.ConfigCache;
+import com.configcat.cache.NullConfigCache;
+import com.configcat.cache.SingleValueCache;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
 import com.configcat.polling.PollingMode;
