+++ src/test/java/com/configcat/ConfigCatClientTest.java
@@ -1,5 +1,6 @@
 package com.configcat;
 
+import com.configcat.cache.FailingCache;
 import com.configcat.evaluation.EvaluationDetails;
 import com.configcat.polling.PollingModes;
 import okhttp3.OkHttpClient;
