+++ src/test/java/com/configcat/cache/FailingCache.java
@@ -1,14 +1,14 @@
-package com.configcat;
+package com.configcat.cache;
 
 public class FailingCache extends ConfigCache {
 
     @Override
-    protected String read(String key) throws Exception {
+    public String read(String key) throws Exception {
         throw new Exception();
     }
 
     @Override
-    protected void write(String key, String value) throws Exception {
+    public void write(String key, String value) throws Exception {
         throw new Exception();
     }
 }
